src/utils/qr.py
import io
import qrcode
from PIL import Image


def generate_qr(data: str, size: int = 200) -> Image.Image:
    """Generate QR code image from string data."""
    qr = qrcode.QRCode(
        version=1,
        error_correction=qrcode.constants.ERROR_CORRECT_L,
        box_size=10,
        border=2,
    )
    qr.add_data(data)
    qr.make(fit=True)
    img = qr.make_image(fill_color="black", back_color="white").convert("RGB")
    return img.resize((size, size), Image.LANCZOS)
