requirements.txt
customtkinter>=5.2.0
qrcode[pil]>=7.4
Pillow>=10.0
pyserial>=3.5
psutil>=5.9
